+++ rasca-api/src/main/java/com/rasca/rascaapi/repositories/UserRepository.java
@@ -3,12 +3,14 @@ package com.rasca.rascaapi.repositories;
 import com.rasca.rascaapi.domain.User;
 import com.rasca.rascaapi.exceptions.EtAuthException;
 
+import java.util.Date;
+
 public interface UserRepository {
-    Integer create(String Correo, String Contrasena, String Usuario, String Nombres, String Apellidos, String Carnet, String FechaNac, String Telefono, String Fotografia) throws EtAuthException;
+    Long create(String Correo, String Contrasena, String Usuario, String Nombres, String Apellidos, String Carnet, String FechaNac, String Telefono, String Fotografia) throws EtAuthException;
 
     User findByEmailAndPassword(String Correo, String Contrasena) throws EtAuthException;
 
-    Integer getCountByEmail(String Correo);
+    Long getCountByEmail(String Correo);
 
-    User findByID(Integer IDPersona);
+    User findByID(Long IDPersona);
 }
